fix third party certificate modal popups
